+++ ysqlite3/vfs/crypt_file.hpp
@@ -8,6 +8,7 @@
 #include <algorithm>
 #include <array>
 #include <cstdint>
+#include <cstdlib>
 #include <cstring>
 #include <openssl/evp.h>
 #include <openssl/rand.h>
@@ -22,7 +23,7 @@ constexpr std::uint8_t crypt_file_reserve_size() noexcept
 }
 
 template<typename File>
-class crypt_file : public vfs::page_transforming_file<File>
+class crypt_file : public page_transforming_file<File>
 {
 public:
 	using page_transforming_file<File>::page_transforming_file;
@@ -30,39 +31,73 @@ public:
 	{
 		EVP_CIPHER_CTX_free(_context);
 	}
-	void file_control(vfs::file_cntl operation, void* arg) override
+	void file_control(file_cntl operation, void* arg) override
 	{
-		if (operation == vfs::file_cntl::pragma) {
+		if (operation == file_cntl::pragma) {
 			auto& name       = static_cast<char**>(arg)[1];
 			const auto value = static_cast<char**>(arg)[2];
 
 			if (!std::strcmp(name, "plain_key")) {
+				auto& key = _transform ? _transformation.key : _current.key;
 				if (!PKCS5_PBKDF2_HMAC(value, -1, nullptr, 0, 200000, EVP_sha512(),
-				                       static_cast<int>(_key.size()), _key.data())) {
-					throw std::system_error{ sqlite3_errc::generic, "failed to generate new key" };
+				                       static_cast<int>(key.size()), key.data())) {
+					name = sqlite3_mprintf("%s", "failed to generate new key");
+				} else {
+					name = sqlite3_mprintf("%s", "ok");
+				}
+			} else if (!std::strcmp(name, "crypt_transformation")) {
+				if ((value[0] == '0' || value[0] == '1') && !value[1]) {
+					if (_transform != (value[0] == '1')) {
+						(_transform ? _current : _transformation) = (_transform ? _transformation : _current);
+						_transform                                = !_transform;
+					}
+					name = sqlite3_mprintf("%s", "ok");
+				} else {
+					name = sqlite3_mprintf("%s", "bad argument");
+				}
+			} else if (!std::strcmp(name, "cipher")) {
+				auto& cipher = _transform ? _transformation.cipher : _current.cipher;
+				if (!std::strcmp(value, "null")) {
+					cipher = nullptr;
+				} else if (!std::strcmp(value, "aes-256-gcm")) {
+					cipher = EVP_aes_256_gcm();
+				} else if (!std::strcmp(value, "aes-192-gcm")) {
+					cipher = EVP_aes_192_gcm();
+				} else if (!std::strcmp(value, "aes-128-gcm")) {
+					cipher = EVP_aes_128_gcm();
+				} else {
+					name = sqlite3_mprintf("%s", "unkown cipher");
+					return;
 				}
 				name = sqlite3_mprintf("%s", "ok");
-				return;
+			} else {
+				goto gt_parent;
 			}
 		}
 
+	gt_parent:
 		page_transforming_file<File>::file_control(operation, arg);
 	}
 
 protected:
 	void encode_page(span<std::uint8_t*> page) override
 	{
-		auto data = page.subspan(page.size() - crypt_file_reserve_size());
-		page      = page.subspan(0, page.size() - crypt_file_reserve_size());
-		_generate_iv(data.subspan(0, 12));
-		_crypt(page, data, true);
-		EVP_CIPHER_CTX_ctrl(_context, EVP_CTRL_GCM_GET_TAG, 16, data.begin() + 12);
+		auto& enc = _transform ? _transformation : _current;
+		if (enc.cipher) {
+			auto data = page.subspan(page.size() - crypt_file_reserve_size());
+			page      = page.subspan(0, page.size() - crypt_file_reserve_size());
+			_generate_iv(data.subspan(0, 12));
+			_crypt(page, data, enc, true);
+			EVP_CIPHER_CTX_ctrl(_context, EVP_CTRL_GCM_GET_TAG, 16, data.begin() + 12);
+		}
 	}
 	void decode_page(span<std::uint8_t*> page) override
 	{
-		auto data = page.subspan(page.size() - crypt_file_reserve_size());
-		page      = page.subspan(0, page.size() - crypt_file_reserve_size());
-		_crypt(page, data, false);
+		if (_current.cipher) {
+			auto data = page.subspan(page.size() - crypt_file_reserve_size());
+			page      = page.subspan(0, page.size() - crypt_file_reserve_size());
+			_crypt(page, data, _current, false);
+		}
 	}
 	bool check_reserve_size(std::uint8_t size) const noexcept override
 	{
@@ -70,15 +105,23 @@ protected:
 	}
 
 private:
-	EVP_CIPHER_CTX* _context  = EVP_CIPHER_CTX_new();
-	const EVP_CIPHER* _cipher = EVP_aes_256_gcm();
-	std::array<unsigned char, 32> _key{};
+	struct encryptor
+	{
+		const EVP_CIPHER* cipher = nullptr;
+		std::array<unsigned char, 32> key{};
+	};
 
-	void _crypt(span<std::uint8_t*> buffer, span<std::uint8_t*> data, bool encrypt)
+	EVP_CIPHER_CTX* _context = EVP_CIPHER_CTX_new();
+	bool _transform          = false;
+	encryptor _current;
+	encryptor _transformation;
+
+	void _crypt(span<std::uint8_t*> buffer, span<std::uint8_t*> data, encryptor& encryptor, bool encrypt)
 	{
 		EVP_CIPHER_CTX_reset(_context);
 
-		if (!EVP_CipherInit_ex(_context, _cipher, nullptr, _key.data(), data.begin(), encrypt)) {
+		if (!EVP_CipherInit_ex(_context, encryptor.cipher, nullptr, encryptor.key.data(), data.begin(),
+		                       encrypt)) {
 			throw std::system_error{ sqlite3_errc::generic, "failed to encrypt/decrypt" };
 		}